Move Curve definition to @haiku/core/api
