+++ packages/@haiku/core/src/StateTransitionManager.ts
@@ -1,12 +1,8 @@
 import Interpolate from './Interpolate';
 import HaikuClock from './HaikuClock';
 import justCurves from './vendor/just-curves';
-import {Curve} from 'haiku-common/lib/types/enums';
-import {
-  CurveFunction,
-  CurveDefinition,
-  BytecodeStateType,
-} from './api/HaikuBytecode';
+import {Curve, CurveFunction, CurveDefinition} from './api/Curve';
+import {BytecodeStateType} from './api/HaikuBytecode';
 
 
 export type StateTransitionParameters = { curve: CurveDefinition, duration: number};
