+++ packages/haiku-common/src/types/enums.ts
@@ -1,37 +1,3 @@
-export enum Curve {
-  EaseInBack = 'easeInBack',
-  EaseInCirc = 'easeInCirc',
-  EaseInCubic = 'easeInCubic',
-  EaseInExpo = 'easeInExpo',
-  EaseInQuad = 'easeInQuad',
-  EaseInQuart = 'easeInQuart',
-  EaseInBounce = 'easeInBounce',
-  EaseInElastic = 'easeInElastic',
-  EaseInQuint = 'easeInQuint',
-  EaseInSine = 'easeInSine',
-  EaseOutBack = 'easeOutBack',
-  EaseOutCirc = 'easeOutCirc',
-  EaseOutCubic = 'easeOutCubic',
-  EaseOutExpo = 'easeOutExpo',
-  EaseOutQuad = 'easeOutQuad',
-  EaseOutQuart = 'easeOutQuart',
-  EaseOutBounce = 'easeOutBounce',
-  EaseOutElastic = 'easeOutElastic',
-  EaseOutQuint = 'easeOutQuint',
-  EaseOutSine = 'easeOutSine',
-  EaseInOutBack = 'easeInOutBack',
-  EaseInOutCirc = 'easeInOutCirc',
-  EaseInOutCubic = 'easeInOutCubic',
-  EaseInOutExpo = 'easeInOutExpo',
-  EaseInOutQuad = 'easeInOutQuad',
-  EaseInOutQuart = 'easeInOutQuart',
-  EaseInOutBounce = 'easeInOutBounce',
-  EaseInOutElastic = 'easeInOutElastic',
-  EaseInOutQuint = 'easeInOutQuint',
-  EaseInOutSine = 'easeInOutSine',
-  Linear = 'linear',
-}
-
 export enum TourUtils {
   ProjectName = 'percy',
 }
