+++ packages/haiku-formats/src/exporters/timelineUtils.ts
@@ -6,7 +6,7 @@ import {
   BytecodeTimelineValue,
   BytecodeTimelineProperty,
 } from '@haiku/core/lib/api/HaikuBytecode';
-import {Curve} from 'haiku-common/lib/types/enums';
+import {Curve, CurveFunction, CurveDefinition} from '@haiku/core/lib/api/Curve';
 
 /**
  * Gets the initial value of a timeline property.
